+++ Ranking.cpp
@@ -78,7 +78,7 @@ int loadData(char *filename, int rank[])
 
 
 
-void drawRankings(int _x, int _y, int ranks[], int _rank_update_info, SPR_DATA* data)
+void drawRankings(float _x, float _y, int ranks[], int _rank_update_info, SPR_DATA* data)
 {
 
 	/////////////////////////////////////////////////////////
@@ -92,7 +92,7 @@ void drawRankings(int _x, int _y, int ranks[], int _rank_update_info, SPR_DATA* 
 	for (int i = 0; i < RANKS_MAX; i++) //5thまで繰り返し
 	{
 		//描画(新しくランクインしたスコアは赤く光る)
-		if (_rank_update_info == i) Render3(V2(_x,_y+i*64-34),&data[0],ranks[i],0xFFFF0000); //赤
+		if (_rank_update_info == i) Render3(V2 (_x,  _y+i*64-34),&data[0],ranks[i],0xFFFF0000); //赤
 		else Render3(V2(_x,_y+i*64-34),&data[0],ranks[i],0xFFFFFFFF); //白
 	}
 	//Render3(V2(_x,_y+(RANKS_MAX+1)*64-34),&data[0],ranks[RANKS_MAX+1],0xFFFF0000); //赤
